Add files via upload
with inheritance
