+++ client.py
@@ -3,13 +3,13 @@
 
 import socket
 import functions
+import basic
 
 
-class Client(object):
+class Client(basic.Basic):
     TCP_IP = "127.0.0.1"    #host
-    TCP_PORT = 5005     #numer portu
+    TCP_PORT = 5011    #numer portu
     BUFFER_SIZE = 1024
-    board = []
     socket = 0
 
     def __init__(self):
@@ -19,46 +19,6 @@ class Client(object):
     def connect(self):
         self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         self.socket.connect((self.TCP_IP, self.TCP_PORT))
-
-    def print_board(self):
-        for field in range(9):
-            if field%3 == 0:
-                print("\n-----------")
-            print(" %c " % self.board[field], end = '')
-            if (field+1)%3 != 0:
-                print("|", end = '')
-        print("\n-----------")
-
-    @staticmethod
-    def mock_board():
-        print("How to play:\n", "Choose where do you want to play your dot or cross,\n",
-              "by typing a number corresponding to the field below")
-        for field in range(9):
-            if field % 3 == 0:
-                print("\n-----------")
-            print(" " + str(field) + " ", end='')
-            if (field + 1) % 3 != 0:
-                print("|", end='')
-        print("\n-----------")
-
-    def check_if_given_field_correct(self, field):
-        if isinstance(field, int):
-            int(field)
-            if 0 <= field <= 8:
-                if self.board[field] == ' ':
-                    print("Your choice was saved.")
-                    return True
-                else:
-                    print("This field is already occupied. Please choose another free value.")
-                    return False
-            else:
-                print("Bad number, use a number from range 0-8 to make a move. For reference:")
-                self.mock_board()
-                return False
-        else:
-            print("Bad value, you have to use integer. Use a number from range 0-8 to make a move. For reference:")
-            self.mock_board()
-            return False
 
     def receive_date(self):
         serialized_board = self.socket.recv(self.BUFFER_SIZE)
